requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.1.3
django-decouple==2.1
django-filter==2.0.0
django-import-export==1.1.0
djangorestframework==3.9.0
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.9
psycopg2==2.7.5
pycodestyle==2.4.0
pylint==2.2.1
pytz==2018.7
PyYAML==3.13
requests==2.20.1
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
typed-ast==1.1.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==3.3.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
